deleted helper.ts file from pages router
